adding custom errorHandling hook
