+++ src/hoc/withErrorHandler/withErrorHandler.js
@@ -1,39 +1,20 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 
 import Modal from "../../components/UI/Modal/Modal";
 import Auxiliary from "../Auxiliary/Auxiliary";
+import useHttpErrorHandler from "../../hooks/http-error-handler";
+
 /**
  *
  * @param {React.Component} WrappedComponent
  */
 const withErrorHandler = (WrappedComponent, axios) => {
   return props => {
-    const [error, setError] = useState(null);
+    const [error, clearError] = useHttpErrorHandler(axios);
 
-    const reqInterceptor = axios.interceptors.request.use(req => {
-      setError(null);
-      return req;
-    });
-    const resInterceptor = axios.interceptors.response.use(
-      res => res,
-      err => {
-        setError(err);
-      }
-    );
-
-    useEffect(() => {
-      return () => {
-        axios.interceptors.request.eject(reqInterceptor);
-        axios.interceptors.response.eject(resInterceptor);
-      };
-    }, [reqInterceptor, resInterceptor]);
-
-    const errorConfirmedHandler = () => {
-      setError(null);
-    };
     return (
       <Auxiliary>
-        <Modal show={error} modalClosed={errorConfirmedHandler}>
+        <Modal show={error} modalClosed={clearError}>
           {error ? error.message : null}
         </Modal>
         <WrappedComponent {...props} />
